refactor: Shared AuthProvider
