+++ src/apps/crm/context/index.js
@@ -1,2 +1,1 @@
-export * from './auth'
 export * from './tabs'
