+++ src/apps/inventory/context/index.js
@@ -4,7 +4,6 @@ import {
    SachetPackagingContext,
 } from './sachetPackaging'
 
-export * from './auth'
 export * from './tabs'
 export * from './item'
 export * from './supplier'
