+++ src/apps/online_store/context/index.js
@@ -1,4 +1,3 @@
-export * from './auth'
 export * from './tabs'
 export * from './collection'
 //  export * from './product'
