+++ src/apps/online_store/index.jsx
@@ -1,14 +1,27 @@
 import React from 'react'
-
-import { AuthProvider, TabProvider } from './context'
+import Keycloak from 'keycloak-js'
 
 import App from './App'
+import { TabProvider } from './context'
+import { AuthProvider } from '../../shared/providers'
+
+const keycloak = new Keycloak({
+   realm: process.env.REACT_APP_KEYCLOAK_REALM,
+   url: process.env.REACT_APP_KEYCLOAK_URL,
+   clientId: 'online-store',
+   'ssl-required': 'none',
+   'public-client': true,
+   'bearer-only': false,
+   'verify-token-audience': true,
+   'use-resource-role-mappings': true,
+   'confidential-port': 0,
+})
 
 const OnlineStore = () => (
-   <AuthProvider>
-   <TabProvider>
-      <App />
-   </TabProvider>
+   <AuthProvider keycloak={keycloak}>
+      <TabProvider>
+         <App />
+      </TabProvider>
    </AuthProvider>
 )
