+++ src/apps/order/context/index.js
@@ -1,4 +1,3 @@
-export * from './auth'
 export * from './tabs'
 export * from './order'
 export * from './access'
